# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Movie Theater API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token to access protected endpoints"
      }
    },
    "schemas": {
      "Genre": {
        "type": "object",
        "properties": {
          "genre_id": {
            "type": "integer"
          },
          "genre_name": {
            "type": "string"
          }
        }
      },
      "Director": {
        "type": "object",
        "properties": {
          "director_id": {
            "type": "integer"
          },
          "director_name": {
            "type": "string"
          }
        }
      },
      "Actor": {
        "type": "object",
        "properties": {
          "actor_id": {
            "type": "integer"
          },
          "actor_name": {
            "type": "string"
          }
        }
      },
      "Showtime": {
        "type": "object",
        "properties": {
          "showtime_id": {
            "type": "integer"
          },
          "movie_id": {
            "type": "integer"
          },
          "room_id": {
            "type": "integer"
          },
          "show_datetime": {
            "type": "string",
            "format": "date",
            "example": "2023-01-01T00:00:00.000Z"
          }
        }
      },
      "MovieDetail": {
        "type": "object",
        "properties": {
          "movie_id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "poster_image": {
            "type": "string",
            "nullable": true
          },
          "poster_url": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "age_rating": {
            "type": "integer",
            "nullable": true
          },
          "run_time": {
            "type": "integer",
            "nullable": true
          },
          "release_date": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "trailer_link": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "type": "string",
            "nullable": true
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genre"
            }
          },
          "director": {
            "$ref": "#/components/schemas/Director"
          },
          "actors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Actor"
            }
          },
          "showtime": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Showtime"
            }
          }
        }
      },
      "MovieUpdateRequest": {
        "type": "object",
        "properties": {
          "poster_image": {
            "type": "string",
            "nullable": true
          },
          "poster_url": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "age_rating": {
            "type": "integer",
            "nullable": true
          },
          "run_time": {
            "type": "integer",
            "nullable": true
          },
          "release_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "trailer_link": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "type": "string",
            "nullable": true
          },
          "director_id": {
            "type": "integer",
            "nullable": true
          }
        }
      },
...
      "Membership": {
        "type": "object",
        "properties": {
          "membership_id": { "type": "integer" },
          "membership_name": { "type": "string" },
          "discount_rate": { "type": "number", "format": "float" }
        }
      },
...
      "Room": {
        "type": "object",
        "properties": {
          "room_id": {
            "type": "integer"
          },
          "room_name": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/movies/{movie_id}": {
      "get": {
        "tags": ["Movie Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "movie_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
         "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieDetail"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Movie Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "movie_id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Movie updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieDetail"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Movie Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "movie_id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Movie deleted"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/rooms/{room_id}": {
      "get": {
        "tags": ["Room Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Room Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "room_name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Room Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Room deleted successfully" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/showtimes/{showtime_id}": {
      "get": {
        "tags": ["Showtime Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "showtime_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Showtime"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Showtime Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "showtime_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Showtime"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Showtime Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "showtime_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Showtime deleted successfully" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/memberships/{membership_id}": {
      "get": {
        "tags": ["Membership Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "membership_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Membership Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "membership_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Membership"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Membership Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "membership_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Membership deleted successfully" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/memberships/create": {
      "post": {
        "tags": ["Membership Controller"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Membership"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership"
                }
              }
            }
          }
        }
      }
    }
  },
...
}